fix hiding hud logic bug and move grammars
fix logic bug with hiding HUD and child views.
move grammars into same folder as voicecommandmanager, maven was having
problems building elsewhere.
