+++ src/main/java/julius/sky/voicehud/App.java
@@ -5,7 +5,7 @@ import com.jme3.asset.AssetManager;
 
 import julius.sky.voicehud.core.hud.HUDGUI;
 import julius.sky.voicehud.core.hud.HUDGUI.GuiLayer;
-import julius.sky.voicehud.core.voice.voicecommand.VoiceCommandManager;
+import julius.sky.voicehud.core.voice.VoiceCommandManager;
 import julius.sky.voicehud.plugins.musicplayer.MusicPlayer;
 
 //public class App extends SimpleApplication
